+++ app/src/main/java/com/s7evensoftware/mobileminer/noso/Constants.kt
@@ -1,7 +1,7 @@
 package com.s7evensoftware.mobileminer.noso
 
 const val SOURCE = "mainnet"
-const val DEFAULT_ADDRESS = "N2kFAtGWLb57Qz91sexZSAnYwA3T7Cy"
+const val DEFAULT_ADDRESS = "NHsx5MnV7UrNubBshnhpSnBuwKaqGw"
 const val DEFAULT_CPUS = 1
 const val DEFAULT_MINING_MODE = true // true = solo mining // false = pool mining
 
@@ -9,6 +9,7 @@ const val NODE_TIMEOUT        = 1500
 const val DELTA_TRIGGER       = true // ff false then nodes with Delta > 0 will be excluded from node selection
 const val DEFAULT_SYNC_DELAY  = 10000L
 const val DEFAULT_MINER_ID    = 0
+const val DEFAULT_POOL_STRING = "45.146.252.103:8082"
 
 const val CoinChar = "N"                   // Char for addresses
 const val B58Alphabet = "123456789ABCDEFGHJKLMNPQRSTUVWXYZabcdefghijkmnopqrstuvwxyz"
@@ -44,6 +45,7 @@ const val MINER_TEST_STOP = "MINERTEST_STOP"
 
 const val EXIT_COMMAND = "EXIT"
 const val MINE_COMMAND = "MINE"
+const val MINE_POOL_COMMAND = "MINEPOOL"
 const val UPDATE_BLOCK_COMMAND = "CHANGEBLOCK"
 const val UPDATE_TARGET_COMMAND = "CHANGETARGET"
 const val SPEEDREPORT_COMMAND = "SPEEDREPORT"
@@ -59,3 +61,4 @@ const val SHAREDPREF_ADDRESS = "MINING_ADDRESS"
 const val SHAREDPREF_CPUS = "MINING_CPUS"
 const val SHAREDPREF_MINERID = "MINING_ID"
 const val SHAREDPREF_MODE = "MINING_MODE"
+const val SHAREDPREF_POOL = "POOL_STRING"