+++ app/src/main/java/com/s7evensoftware/mobileminer/noso/Fragments/SettingsFragment.kt
@@ -22,6 +22,7 @@ class SettingsFragment : Fragment(), View.OnClickListener {
     private lateinit var binding:FragmentSettingsBinding
     private lateinit var viewModel: MainViewModel
     private lateinit var callback:SettingsFragmentListener
+    private var localMode = false
 
     companion object {
         fun newInstance() = SettingsFragment()
@@ -50,18 +51,40 @@ class SettingsFragment : Fragment(), View.OnClickListener {
         binding.settingsFragmentAddressPaste.setOnClickListener(this)
         binding.settingsFragmentCpuSlider.valueTo = viewModel.CPUcores.toFloat()
         binding.settingsFragmentMinerID.value = viewModel.MinerID
+        binding.settingsFragmentMode.setOnClickListener(this)
 
         binding.settingsFragmentCpuSlider.addOnChangeListener { _, value, _ ->
             binding.settingsFragmentCpuNumber.text = value.toInt().toString()
         }
 
-        viewModel.CPUtoUse.observe(viewLifecycleOwner, { number ->
+        viewModel.CPUtoUse.observe(viewLifecycleOwner) { number ->
             binding.settingsFragmentCpuSlider.value = number.toFloat()
-        })
+        }
 
-        viewModel.MinerAddress.observe(viewLifecycleOwner, { address ->
+        viewModel.MinerAddress.observe(viewLifecycleOwner) { address ->
             binding.settingsFragmentAddress.setText(address)
-        })
+        }
+
+        viewModel.poolString.observe(viewLifecycleOwner) { poolString ->
+            binding.settingsFragmentPoolSource.setText(poolString)
+        }
+
+        viewModel.isSoloMining.observe(viewLifecycleOwner) { solo ->
+            if(solo){
+                localMode = solo
+                binding.settingsFragmentMode.isChecked = !solo
+                binding.settingsFragmentSourceContainer.visibility = View.GONE
+                binding.settingsFragmentMineridContainer.visibility = View.VISIBLE
+            }else{
+                localMode = solo
+                binding.settingsFragmentMode.isChecked = !solo
+                binding.settingsFragmentSourceContainer.visibility = View.VISIBLE
+                binding.settingsFragmentMineridContainer.visibility = View.GONE
+            }
+        }
+
+
+
 
     }
 
@@ -74,6 +97,18 @@ class SettingsFragment : Fragment(), View.OnClickListener {
         when(v.id){
             R.id.settings_fragment_address_qr -> {
                 callback.onQRScann()
+            }
+            R.id.settings_fragment_mode -> {
+                localMode = !binding.settingsFragmentMode.isChecked
+                if(localMode){
+                    binding.settingsFragmentSourceContainer.visibility = View.GONE
+                    binding.settingsFragmentMineridContainer.visibility = View.VISIBLE
+                }else{
+                    binding.settingsFragmentMode.isChecked = !localMode
+                    binding.settingsFragmentSourceContainer.visibility = View.VISIBLE
+                    binding.settingsFragmentMineridContainer.visibility = View.GONE
+                }
+                //viewModel.isSoloMining.value = !binding.settingsFragmentMode.isChecked
             }
             R.id.settings_fragment_address_paste -> {
                 val clipboard = requireActivity().getSystemService(CLIPBOARD_SERVICE) as ClipboardManager
@@ -89,7 +124,7 @@ class SettingsFragment : Fragment(), View.OnClickListener {
                 viewModel.MinerAddress.value = binding.settingsFragmentAddress.text.toString()
                 viewModel.CPUtoUse.value = binding.settingsFragmentCpuSlider.value.toInt()
                 viewModel.MinerID = binding.settingsFragmentMinerID.value
-                viewModel.isSoloMining = !binding.settingsFragmentMode.isChecked
+                viewModel.isSoloMining.value = !binding.settingsFragmentMode.isChecked
 
                 if(viewModel.getMinerAddres().isNotEmpty() && viewModel.getMinerAddres().isNotBlank() && Nosocoreunit.IsValidAddress(viewModel.getMinerAddres())){
                     callback.onSaveSettings()